requirements.txt
fastapi==0.85.1
google_api_python_client==2.65.0
google_auth_oauthlib==0.4.2
loguru==0.5.3
pandas==1.3.0
uvicorn==0.15.0
python-multipart==0.0.5
python-dotenv==0.19.0
quarter_lib==0.0.6